# artifacts/integration/day04/frontend_openapi.json
{"openapi":"3.1.0","info":{"title":"AI Commerce Operations Agent","version":"0.1.0"},"paths":{"/api/v1/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Api V1 Health Get"}}}}}}},"/api/v1/ready":{"get":{"tags":["meta"],"summary":"Ready","operationId":"ready_api_v1_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Ready Api V1 Ready Get"}}}}}}},"/api/v1/meta":{"get":{"tags":["meta"],"summary":"Meta","operationId":"meta_api_v1_meta_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Meta Api V1 Meta Get"}}}}}}},"/api/v1/demo/events":{"post":{"tags":["demo"],"summary":"Publish Demo Event","operationId":"publish_demo_event_api_v1_demo_events_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CanonicalCommerceEvent"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Publish Demo Event Api V1 Demo Events Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/insights":{"get":{"tags":["demo"],"summary":"Insights","operationId":"insights_api_v1_insights_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiEnvelope_list_dict_str__Any___"}}}}}}}},"components":{"schemas":{"ApiEnvelope_list_dict_str__Any___":{"properties":{"schema_version":{"type":"string","title":"Schema Version","default":"1.0"},"tenant_id":{"type":"string","title":"Tenant Id"},"request_id":{"type":"string","title":"Request Id"},"trace_id":{"type":"string","title":"Trace Id"},"data":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Data"},"evidence_ids":{"items":{"type":"string"},"type":"array","title":"Evidence Ids"},"warnings":{"items":{"type":"string"},"type":"array","title":"Warnings"},"as_of":{"type":"string","format":"date-time","title":"As Of"}},"additionalProperties":false,"type":"object","required":["tenant_id","request_id","trace_id","data","as_of"],"title":"ApiEnvelope[list[dict[str, Any]]]"},"CanonicalCommerceEvent":{"properties":{"event_id":{"type":"string","maxLength":200,"minLength":1,"title":"Event Id"},"event_type":{"$ref":"#/components/schemas/EventType"},"schema_version":{"type":"string","pattern":"^1(?:\\.\\d+)?$","title":"Schema Version"},"tenant_id":{"type":"string","maxLength":100,"minLength":1,"title":"Tenant Id"},"source":{"$ref":"#/components/schemas/Provider"},"source_event_id":{"type":"string","maxLength":255,"minLength":1,"title":"Source Event Id"},"occurred_at":{"type":"string","format":"date-time","title":"Occurred At"},"ingested_at":{"type":"string","format":"date-time","title":"Ingested At"},"idempotency_key":{"type":"string","maxLength":500,"minLength":1,"title":"Idempotency Key"},"correlation_id":{"type":"string","maxLength":200,"minLength":1,"title":"Correlation Id"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"}},"additionalProperties":false,"type":"object","required":["event_id","event_type","schema_version","tenant_id","source","source_event_id","occurred_at","ingested_at","idempotency_key","correlation_id","payload"],"title":"CanonicalCommerceEvent","description":"Immutable fact envelope defined by 05 Data/API specification section 5."},"EventType":{"type":"string","enum":["order.created","order.updated","offline_sale.recorded","return.recorded","inventory.snapshot_received","expected_changed","discrepancy_detected","reservation.created","reservation.shortage_detected","incoming.updated","inquiry.received","incoming.delayed","action.denied"],"title":"EventType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Provider":{"type":"string","enum":["CAFE24","TOSS_POS","ECOUNT","DEMO"],"title":"Provider"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}